Prep for Site Water content

+++ Classes/Arch_Classes/SurfaceCrew.m
@@ -71,6 +71,6 @@ classdef SurfaceCrew < handle
         MIN_CREW = SurfaceCrew(double(2));
         DEFAULT_TRANSIT = SurfaceCrew(double(4));
         DRA_CREW = SurfaceCrew(double(6));
-        TARGET_SURFACE = SurfaceCrew(double(20));
+        TARGET_SURFACE = SurfaceCrew(double(24));
     end
 end
